Put negative_sampling methods into utils module.
